Moved everything from compute/kernels/impl to medium
